Save reference to native bind apply method
Saved a reference to the native behavior: `Function.prototype.bind.apply( ... )`.
The extension's tracing code uses this native logic to track the construction of AudioContexts and AudioNodes. A web page using a library that overrides the native bind and apply methods may interfere with this tracking without this change.

For instance, the famous box demo used to be broken when Web Audio Inspector is enabled because it uses prototype.js:
http://googlechrome.github.io/web-audio-samples/samples/audio/box2d-js/box2d-audio.html